requirements.txt
# Core PyTorch and ML dependencies
torch>=2.0.0
torchvision
transformers>=4.21.0
tokenizers>=0.19.1
einops>=0.8.0
numpy>=1.26.4
safetensors

# Vision and Image Processing
opencv-python>=4.6.0.66
imageio>=2.34.1
matplotlib>=3.9.0
Pillow

# RIPT Dependencies for Reinforcement Learning Training
hydra-core==1.3.2
bddl==1.0.1
robosuite==1.4.1
gym==0.25.2
cloudpickle==2.1.0
pyinstrument==4.6.2
moviepy==1.0.3

# Additional RIPT utilities
easydict==1.13
h5py==3.11.0
positional-encodings==6.0.3
tqdm>=4.66.4
vector-quantize-pytorch==1.14.24
wandb>=0.18.3
cmake>=3.29.3
diffusers>=0.28.0
torchtyping==0.1.4
ema-pytorch==0.5.2
natsort==8.4.0
thop==0.1.1-2209072238
future==0.18.2

# Development and testing
pytest
jupyter

# 数据处理与科学计算
pandas>=2.0.0
scipy>=1.10.0

# 机器学习与图像处理
scikit-learn>=1.3.0
scikit-image>=0.24.0

# YAML 配置
PyYAML